Refactor: Move existing Schedule views into v1 subfolder
