+++ frontend/src/aspects/Conference/Attend/Schedule/v1/NowMarker.tsx
@@ -1,5 +1,5 @@
 import React, { useCallback, useEffect, useRef } from "react";
-import { useRealTime } from "../../../Generic/useRealTime";
+import { useRealTime } from "../../../../Generic/useRealTime";
 import useTimelineParameters from "./useTimelineParameters";
 
 export default function NowMarker({
