+++ frontend/src/aspects/Conference/Attend/Schedule/v1/Scaling.tsx
