+++ frontend/src/aspects/Conference/Attend/Schedule/v1/Schedule.tsx
@@ -14,12 +14,12 @@ import {
     Schedule_SelectSummariesQuery,
     Schedule_TagFragment,
     useSchedule_SelectSummariesQuery,
-} from "../../../../generated/graphql";
-import ApolloQueryWrapper from "../../../GQL/ApolloQueryWrapper";
-import { FAIcon } from "../../../Icons/FAIcon";
-import { useTitle } from "../../../Utils/useTitle";
-import RequireAtLeastOnePermissionWrapper from "../../RequireAtLeastOnePermissionWrapper";
-import { useConference } from "../../useConference";
+} from "../../../../../generated/graphql";
+import ApolloQueryWrapper from "../../../../GQL/ApolloQueryWrapper";
+import { FAIcon } from "../../../../Icons/FAIcon";
+import { useTitle } from "../../../../Utils/useTitle";
+import RequireAtLeastOnePermissionWrapper from "../../../RequireAtLeastOnePermissionWrapper";
+import { useConference } from "../../../useConference";
 import DayList, { TimelineEvent } from "./DayList";
 import DownloadCalendarButton from "./DownloadCalendarButton";
 import NowMarker from "./NowMarker";
@@ -367,10 +367,10 @@ function ScheduleFrame({
     );
 
     const timelineParams = useTimelineParameters();
-    const startTime = useMemo(() => DateTime.fromMillis(frame.startTimeMs).setZone(timelineParams.timezone), [
-        frame.startTimeMs,
-        timelineParams.timezone,
-    ]);
+    const startTime = useMemo(
+        () => DateTime.fromMillis(frame.startTimeMs).setZone(timelineParams.timezone),
+        [frame.startTimeMs, timelineParams.timezone]
+    );
     const borderColourRaw = useToken("colors", borderColour);
     const { colorMode } = useColorMode();
 
@@ -509,9 +509,10 @@ export function ScheduleInner({
 
     const [scrollToNow, setScrollToNow] = useState<{ f: () => void } | null>(null);
 
-    const maxParallelRooms = useMemo(() => frames.reduce((acc, frame) => Math.max(acc, frame.items.length), 0), [
-        frames,
-    ]);
+    const maxParallelRooms = useMemo(
+        () => frames.reduce((acc, frame) => Math.max(acc, frame.items.length), 0),
+        [frames]
+    );
     const timeBarWidth = 50;
     const roomColWidth = Math.min(
         800,
