+++ frontend/src/aspects/Conference/Attend/Schedule/v1/TimeBar.tsx
